# Conflicts:
#	FrontEnd/majorproject-front-end/src/App.js

+++ BackEnd/src/main/java/septmajorproject/bookingsys/repository/CustomerRepository.java
@@ -1,16 +1,15 @@
 package septmajorproject.bookingsys.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import septmajorproject.bookingsys.model.Customer;
 
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     @Override
-    //returns a list of all customers
+        //returns a list of all customers
     List<Customer> findAll();
 
     //finds customer by unique username
